Make infeasibily more changeable : remove old class Constraint (now useless) + move a header to the right location
