+++ src/solver/infeasible-problem-analysis/CMakeLists.txt
@@ -10,9 +10,7 @@ set(SRC_INFEASIBLE_PROBLEM_ANALYSIS
         include/antares/solver/infeasible-problem-analysis/unfeasible-pb-analyzer.h
         include/antares/solver/infeasible-problem-analysis/report.h
         report.cpp
-        include/antares/solver/infeasible-problem-analysis/constraint.h
-        constraint.cpp
-        detected-infeasible-constraints.h
+        include/antares/solver/infeasible-problem-analysis/detected-infeasible-constraints.h
         detected-infeasible-constraints.cpp
 )
